+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 * [artifact-manager] 148 / vra-ng:pull command failure when import/export tags contains underscore
 * [maven-plugins] 146 / Pull operations fail on Windows with release 2.34.0
+* [artifact-manager] 778 / Backup only works with current package version, otherwise throws exception 404 not found
 
 ## v2.34.0 - 05 Jul 2023